fixed all battle elements, added beyblade evade

+++ utils.py
@@ -4,4 +4,4 @@ import math
 def calc_distance(ax, ay, bx, by):
     dx = bx-ax
     dy = by-ay
-    return math.sqrt(math.pow(dx, 2) + math.pow(dy, 2))
+    return int(math.sqrt(math.pow(dx, 2) + math.pow(dy, 2)))
